# openapi.yaml
openapi: 3.0.3
info:
  title: Cling AI Go 用户端 API
  version: 0.1.0-local
  description: 本文档只描述本地 Go 用户端合同；管理后台、Animate、真实支付供应商和生产地址不在范围内。
servers:
  - url: http://127.0.0.1:18000
    description: 本地 Gateway
security:
  - bearerAuth: []
paths:
...
  /api/chat/image/async:
    post:
      summary: 创建 T2I 或模板 I2I 任务
      description: 门禁、预扣、审核和幂等由 Go 服务端处理，不接受余额、VIP 或钻石字段。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageCreationRequest' }
      responses:
        '200': { $ref: '#/components/responses/CreationResponse' }
        '402': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
  /api/chat/video:
    post:
      summary: 创建模板视频任务
      description: 文生视频由 Go 编排为首帧和图生视频两个步骤；Animate 不属于此接口。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VideoCreationRequest' }
      responses:
        '200': { $ref: '#/components/responses/CreationResponse' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
...
    ImageCreationRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string }
        templateId: { type: string }
        assetIds: { type: array, items: { type: string } }
        idempotencyKey: { type: string }
    VideoCreationRequest:
      type: object
      required: [templateId, durationSeconds]
      properties:
        templateId: { type: string }
        prompt: { type: string }
        imageUrl: { type: string, format: uri }
        durationSeconds: { type: integer, enum: [5, 10, 15] }
        idempotencyKey: { type: string }
...
    Error:
      type: object
      properties:
        success: { type: boolean, enum: [false] }
        code: { type: string }
        message: { type: string }
  responses:
...
    CreationResponse: { description: 创建结果, content: { application/json: { schema: { type: object } } } }
...
    ErrorResponse: { description: 业务错误, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }